feat: introduce addon system
