+++ src/types.ts
@@ -1,3 +1,5 @@
+/* eslint-disable no-use-before-define */
+import MagicString from 'magic-string'
 import type { BuiltinPresetName } from './presets'
 
 export type ModuleId = string
@@ -28,10 +30,28 @@ export interface Preset extends ImportCommon {
   imports: (PresetImport | Preset)[]
 }
 
+export interface UnimportContext {
+  readonly imports: Import[]
+  staticImports: Import[]
+  dynamicImports: Import[]
+  map: Map<string, Import>
+  addons: Addon[]
+}
+
+export interface AddonsOptions {
+  /**
+   * Enable auto import inside for Vue's <template>
+   *
+   * @default false
+   */
+  vueTemplate?: boolean
+}
+
 export interface UnimportOptions {
   imports: Import[]
   presets: (Preset | BuiltinPresetName)[]
   warn: (msg: string) => void
+  addons: AddonsOptions
 }
 
 export type PathFromResolver = (_import: Import) => string | undefined
@@ -58,9 +78,9 @@ export interface InjectImportsOptions {
    * @default false
    */
   mergeExisting?: boolean
+}
 
-  /**
-   * @default false
-   */
-  vueTemplate?: boolean
+export interface Addon {
+  transform: (code: string | MagicString, ctx: UnimportContext) => MagicString | Promise<MagicString>
+  decleration?: (dts: string) => string
 }
